+++ app/Http/Controllers/DriversController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\{Drivers,TradeMarks, Routes, Cars, CarsExits};
 use App\Http\Requests\Drivers\{CreateRequest,UpdateRequest};
-use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\{DB, Gate};
 use Illuminate\Auth\Middleware\Authenticate;
 use App\Events\DeleteAudit;
 use Barryvdh\DomPDF\Facade as PDF;
@@ -22,6 +22,9 @@ class DriversController extends Controller
     }
     public function index()
     {
+        if (! Gate::allows('drivers.index')) {
+            return abort(401);
+        }
     	return view('drivers.index',[
     		'drivers'=> Drivers::all(),
         ]);
@@ -46,6 +49,9 @@ class DriversController extends Controller
 
     public function create()
     {
+        if (! Gate::allows('drivers.create')) {
+            return abort(401);
+        }
     	return view('drivers.create',[
             'drivers' => new drivers,
         ]);
@@ -53,6 +59,9 @@ class DriversController extends Controller
 
     public function store(CreateRequest $request)
     {
+        if (! Gate::allows('drivers.create')) {
+            return abort(401);
+        }
     	$request->driversCreate();
 
     	return redirect()->route('drivers.index',[
@@ -62,6 +71,9 @@ class DriversController extends Controller
 
     public function show(Drivers $driver)
     {
+        if (! Gate::allows('drivers.edit')) {
+            return abort(401);
+        }
         return view('drivers.edit',[
             'drivers' => $driver,
         ]);
@@ -69,6 +81,9 @@ class DriversController extends Controller
 
     public function update(UpdateRequest $request, Drivers $driver)
     {
+        if (! Gate::allows('drivers.edit')) {
+            return abort(401);
+        }
         $request->driversUpdate($driver);
 
         return redirect()->route('drivers.index',[
@@ -78,6 +93,9 @@ class DriversController extends Controller
 
       public function delete(Drivers $driver)
     {
+        if (! Gate::allows('drivers.deleteSoft')) {
+            return abort(401);
+        }
         $driver->delete();
 
         return redirect()->route('drivers.index',[
@@ -87,6 +105,9 @@ class DriversController extends Controller
 
     public function restore(Request $request)
     {
+        if (! Gate::allows('drivers.restore')) {
+            return abort(401);
+        }
        $driver = Drivers::onlyTrashed()->findOrFail($request->id);
        $driver->restore();
 
@@ -97,12 +118,18 @@ class DriversController extends Controller
 
     public function onlyTrashed()
     {
+        if (! Gate::allows('drivers.onlyTrashed')) {
+            return abort(401);
+        }
         return view('drivers.onlyTrashed',[
             'drivers' => Drivers::onlyTrashed()->get()
         ]);
     }
      public function remove(Request $request)
     {
+        if (! Gate::allows('drivers.destroy')) {
+            return abort(401);
+        }
         Drivers::where('id', $request->id)->forceDelete();
        
         return view('drivers.onlyTrashed',[
