+++ app/Http/Controllers/EmployeesController.php
@@ -5,18 +5,25 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Employees\{CreateRequest, UpdateRequest};
 use App\Models\Employee;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\{DB, Gate};
 
 
 class EmployeesController extends Controller
 {
     public function index()
     {
+        if (! Gate::allows('employees.index')) {
+            return abort(401);
+        }
         return view('employees.index', [
             'employees' => Employee::all()
         ]);
     }
     public function create()
     {
+        if (! Gate::allows('employees.create')) {
+            return abort(401);
+        }
         return view('employees.create', [
             'employees' => new Employee
         ]);
@@ -24,6 +31,9 @@ class EmployeesController extends Controller
 
     public function store(CreateRequest $request)
     {
+        if (! Gate::allows('employees.create')) {
+            return abort(401);
+        }
         $request->EmployeeCreate();
 
         return view('employees.index', [
@@ -33,12 +43,18 @@ class EmployeesController extends Controller
 
     public function edit(Employee $empleado)
     {
+        if (! Gate::allows('employees.edit')) {
+            return abort(401);
+        }
         return view('employees.edit', [
             'employees' =>  $empleado
         ]);
     }
     public function destroy(Employee $empleado)
     {
+        if (! Gate::allows('employees.deleteSoft')) {
+            return abort(401);
+        }
         $empleado->delete();
 
         return redirect()->route('empleados.index', [
@@ -47,6 +63,9 @@ class EmployeesController extends Controller
     }
     public function remove(Request $request)
     {
+        if (! Gate::allows('employees.softDelete')) {
+            return abort(401);
+        }
         Employee::where('id', $request->id)->forceDelete();
        
         return redirect()->route('empleados.onlyTrashed', [
@@ -56,6 +75,9 @@ class EmployeesController extends Controller
 
     public function update(UpdateRequest $request,  Employee $empleado)
     {
+        if (! Gate::allows('employees.edit')) {
+            return abort(401);
+        }
         $request->employeeUpdate($empleado);
 
         return view('employees.index', [
@@ -65,6 +87,9 @@ class EmployeesController extends Controller
 
     public function onlyTrashed()
     {
+        if (! Gate::allows('employees.onlyTrashed')) {
+            return abort(401);
+        }
         return view('employees.onlyTrashed', [
             'employees' => Employee::onlyTrashed()->get()
         ]);
@@ -72,6 +97,10 @@ class EmployeesController extends Controller
 
     public function restore(Request $request)
     {
+        
+        if (! Gate::allows('employees.restore')) {
+            return abort(401);
+        }
         $employee = Employee::onlyTrashed()->findOrFail($request->id);
         $employee->restore();
