+++ app/Http/Controllers/StopsController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\{Addreses,Stops};
 use App\Http\Requests\Stops\{CreateRequest};
-
+use Illuminate\Support\Facades\{DB, Gate};
 
 class StopsController extends Controller
 {
@@ -19,6 +19,10 @@ class StopsController extends Controller
 
     public function toggleStatus(Stops $stop)
     {
+        if (! Gate::allows('top.toggleStatus')) {
+            return abort(401);
+        }
+        
     	$address = $stop->addreses;
     	$stop->status = ($stop->status == 0)? 1 : 0;
         $stop->update();
@@ -31,6 +35,9 @@ class StopsController extends Controller
 
     public function store(CreateRequest $request, Addreses $address)
     {
+        if (! Gate::allows('top.create')) {
+            return abort(401);
+        }
 		$request->createStops($address);
 
     	return redirect()->route('stops.index',[
@@ -41,6 +48,9 @@ class StopsController extends Controller
 
     public function destroy(Stops $stop)
     {
+        if (! Gate::allows('top.destroy')) {
+            return abort(401);
+        }
         $address = $stop->Addreses;
         $stop->delete();
         return view('stops.index',[
